+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## unreleased version -- v0.11.3
 ### Added
+- New endpoint to update whole collection JSON objects in the API response
+### Fixed
 - Add new DB entities for automatic collections
 - Add support for 'Most Recent' automatic collection
-### Fixed
 
 ## 2023-02-02 -- v0.11.2
 ### Added